Python - Variable Annotations. Task 2. Basic annotations - floor. fixing return type

+++ python_variable_annotations/2-floor.py
@@ -4,4 +4,4 @@
 
 def floor(n: float) -> float:
     """Task 2"""
-    return floor(n)
+    return float(floor(n))
